refactor: rename proyectoByInvestigadorFolder
