+++ src/app/ui/index.ts
@@ -3,5 +3,5 @@ export * from "./button"
 export * from "./cards"
 export * from "./form"
 export * from "./Pagination"
+export * from "./proyectoByInvestigador/ProyectoViewerByInvestigador"
 export * from "./Search"
-export * from "./viewer/ProyectoViewerByInvestigador"
