+++ src/app/ui/proyectoByInvestigador/ProyectoViewerByInvestigador.tsx
